requirements.txt
django>=5.0,<5.1
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3.1,<4.4
psycopg2>=2.9,<2.10
drf-spectacular>=0.27,<0.28
openai>=1.13,<1.14
langchain>=0.1,<0.2
langchain-openai>=0.1.3,<0.2
django-storages>=1.14,<1.15
boto3>=1.34,<1.35
uwsgi>=2.0.25.1,<2.0.26
sendgrid>=6.11.0,<6.12